+++ deslab/__init__.py
@@ -61,8 +61,9 @@ from deslab.toolboxes.supervisory import *
 #from deslab.toolboxes.supervisory import *
 from deslab.toolboxes.opacity_verifier import *
 from deslab.toolboxes.opacity_enforcement import *
-from deslab.toolboxes.opacity_TIA import *
+from deslab.toolboxes.ti_functions import *
 from deslab.toolboxes.ti_diagnosis import *
+from deslab.toolboxes.ti_opacity_verifier import *
 
 # adding readwrite
 import deslab.readwrite
