+++ deslab/toolboxes/__init__.py
@@ -38,6 +38,9 @@ Using
 
 from deslab.toolboxes.diagnosis import *
 from deslab.toolboxes.supervisory import *
+#from deslab.toolboxes.supervisory import *
 from deslab.toolboxes.opacity_verifier import *
 from deslab.toolboxes.opacity_enforcement import *
-from deslab.toolboxes.opacity_TIA import *
+from deslab.toolboxes.ti_functions import *
+from deslab.toolboxes.ti_diagnosis import *
+from deslab.toolboxes.ti_opacity_verifier import *
